its_me/requirements.txt
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==2.2.19
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.0
yarl==1.9.4
